Add function to poll sensors on SKARAB motherboard
Implemented function to invoke the GetSensorData command on the uBlaze.
This function creates the required request packet and handles the
returning response packet, returning the sensor information in a
dictionary.

Added new classes to handle new request/response packet structures.
Added new support for the newly implemented opcode on uBlaze.

Currently the only sensors supported are the motherboard fan speeds.