Fix voltage/current read data handling
Re-implemented the support for reading from voltage and current sensors
on the SKARAB. The SKARAB returns raw data that needs to be processed to
extract the voltage or current values.

Support for all the voltage and current sensors has been added.
Additional support for the 3V3 config sensor is required to support
varying board revisions.