Add function to allow reading of HMC registers
Added a new function, read_hmc_i2c, to allow reading of HMC registers via
the I2C interface. This involves invoking a new command on the
microblaze. Added new packet format for this command.

It should be noted that the opcodes for GetSensorData and SetFanSpeed
have been changed in this commit. This was done to create an offset
between Peralex microblaze commands and SKA SA microblaze commands. The
appropriate version of mlib_devel contains this change in the SKARAB
board support package.