Implement support for bulk writes for SKARAB
Can now write > 32b per write transaction. Maximum of 1988 bytes per
transaction, but an arbitrary number of bytes can be written - the write
function will perform as many writes as is required.

Also remove checking for golden, and multiboot images in the
get_embedded_software_version method.

It is now also ensured that the 'prog_info' attribute is updated even
when programming with non fpg files.