Enable setFreq method in synth.py. Remove LMX2581 configuration files. Modify snapadc.py in order to handle signed integer from ADCs Add 12-bit ADC HMCAD1520 class in adc.py and corresponding support in snapadc.py Fix bugs
Demonstration of the way of using synth.py and snapadc.py can be found at https://github.com/ianmalcolm/HERADev/blob/adc_devel/adc16_test.py, in which parameter "snap" is an instance of SNAPADC class, and "lmx" is an instance of LMX2581 class. Currently the failure rate of alignLineClock() and alignFrameClock() is below 1%. The clock aligning test can be reproduced by using the clockAlignmentTest() in adc16_test.py.