Adding support for multiple packet chunk sizes
The programming chunk size used to be hardcoded at compile time in progska, it
is now a parameter in upload_to_ram_and_program. This is mostly due to older
bitstreams not supporting the newer larger packet sizes. Now it defaults to 1988.
This will be changed to 7952 at a later stage once, we have upgraded all the
SKARABs with the new board support firmware.