Add redis transport class
This emulates the tapcp transport but funnels TFTP messages
through a redis PUBSUB connection. This (hopefully) allows multiple users
to talk to SNAP boards without having issues with multiple tftp links to the
SNAP's microblaze.
It also means that devices not on the SNAP-net can talk to the boards, provided
they have access to the redis gateway