Add a method to check that all ADC lanes are in sync
Empirical evidence suggests that sometimes the SNAP ADC(s)
will initialize such that one of the cores is a sample offset
from the others.
This commit adds a test **not automatically checked** so that
the user can quickly see if all cores are aligned.

Strictly, for normal functionality (i.e. "good" spectra)
one doesn't need _all_ the cores to be mutually aligned,
just cores operating together as a single interleaved sampler.
The implemented test does the easiest thing, and makes sure
everything is aligned whether it needs to be or not. This is
probably a pleasant feature anyway.