Can now talk to SKARABs properly.
The main change has been in skarab_definitions.py, the
create_payload() function - previously in p2, a string was
just a set of ASCII, representable in memory as you see it.
Now with p3's unicode-by-default, you have to encode and
unpack (using struct) to get the equivalent bytes.

Also, dict.iteritems() becomes dict.items(), and changed
the string formatting to be p3 version in a few more
places. There are still exceptions that have e.message in
them though, that will need to be changed before the
port can be considered complete.