Add support for reading SKARAB sensor status
JIRA: ST-199

GetSensorData now parses the status bit for each sensor to determine if
the sensor read was successful or not. In the case of a read being
unsuccessful, the returned sensor data is not processed and the sensor
that is in error is set to -1 with the status 'unknown'.