Fix wait_after_reboot function for SKARABs
The wait_after_reboot function waits for SKARABs to come back after
being reprogrammed and checks if they're reachable and running the
correct firmware. Exceptions were not being handled properly in some of
the underlying functions causing the function to bail when not getting a
response from a SKARAB. This has been fixed in this commit.