Clean up upload_to_ram_and_program method
Cleaned up the upload_to_ram_and_program method. If programming a board
fails, we no longer populate the system information. Also cleaned up the
passing of unique parameters to the transport layers. e.g. the
chunk_size parameter for skarab. This can be passed as a kwarg to
upload_to_ram_and_program and is handled appropriately in the underlying
transport layer.