Handle out-of-memory range errors for wishbone r/w
Add error handling for reads and writes to wishbone addresses that are
out of range. Initial support for error handling added. Modified
wishbone command packet structures to include field for error_status.
Maintains backwards compatibility with older MicroBlaze versions. Needs
to be tested. And other methods of transport_skarab need to be
refactored to take full advantage of this.