tftpy sessions are not ended promptly
E.g. Immediately after a successful download from the fpga, read() returns success but it does not call end() to stop the session.  If the fpga is a lot slower than the ftp client (probably) and the ftp client tries a 2nd download quickly, the fpga might think it is still busy from the 1st session.