Fix issues encountered when trying to install casperfpga
The setup.py file was defining progska as a Python module and a Python
extension. As a result of this, katversion, a Python package that
handles version information, was expecting an __init__.py file for
progska. The setup.py file has been modified to treat progska only as a
Python extension and not a module.

Resolves: ST-227