Allow configuration of multicast on all interfaces
Any of the four 40GbE network interfaces on the SKARAB can now be
configured for multicast. Also includes error handling in the event that
the user attempts to configure an interface that is out of range or an
interface that is not present in the compiled design. The IGMP leave
command now also works for any of the 40GbE network interfaces.

Resolves: ST-156