first attempt at brining functionality to casperfpga for rfsoc support
implements the methods to interface with the added katcp methods for
working with rfsoc. This is what was demoed for the casper workshop the only
thing I can really remember to go back and work through was to make a decision
on the katcp upload method to use. Here it looks like I opt'd for my own
implementation of a katcp upload and now need to revisit to see if this is the
long term solution